+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Fixed issue with `FilterPerformance` where the arg `measure` wasn't passed on
 - Added `FilterUnivariateCox` (thanks to @bblodfon)
+- Parameter value `na.rm` is properly initialized to `TRUE` (thanks to @bblodfon)
+- Bugfix: property `missings` is now set correctly for `FilterFindCorrelation`
 
 # mlr3filters 0.7.1